Moving all code generators to one file
